+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -43,6 +43,7 @@ public protocol HIGTheme: Sendable {
     var imageOverlay: any HIGImageOverlayTokens { get }
     var colorSelector: any HIGColorSelectorTokens { get }
     var socialButton: any HIGSocialButtonTokens { get }
+    var imageFrame: any HIGImageFrameTokens { get }
     var dataTable: any HIGDataTableTokens { get }
     var breadcrumb: any HIGBreadcrumbTokens { get }
     var pageHeader: any HIGPageHeaderTokens { get }