+++ Sources/HIGThemesContract/Preview/HIGComponentPreviewTheme.swift
@@ -44,6 +44,7 @@ public struct HIGComponentPreviewTheme: HIGTheme {
     public let imageOverlay: any HIGImageOverlayTokens = HIGSystemImageOverlayTokens()
     public let colorSelector: any HIGColorSelectorTokens = HIGSystemColorSelectorTokens()
     public let socialButton: any HIGSocialButtonTokens = HIGSystemSocialButtonTokens()
+    public let imageFrame: any HIGImageFrameTokens = HIGSystemImageFrameTokens()
     public let dataTable: any HIGDataTableTokens = HIGSystemDataTableTokens()
     public let breadcrumb: any HIGBreadcrumbTokens = HIGSystemBreadcrumbTokens()
     public let pageHeader: any HIGPageHeaderTokens = HIGSystemPageHeaderTokens()